clu.metrics: Fix type annotations for `Metric`
Similar to the previous change for `Collection`, this allows the
methods and classmethods on `Metric` and its subclasses to be properly
typed. This allows us to remove the `disable=g-bare-generic` lint.

This also enables typechecking for `metrics_test.py`, which failed to
typecheck before this change.

Finally, this change uses `Protocol` instead of `abc.ABC` for `Value`,
so that we can call `v.value` on abstract `Value` types.

PiperOrigin-RevId: 527250992